requirements.txt
python-dotenv
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
sqlite3  # Built into Python
python-dotenv>=0.19.0
schedule>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0

# For ML models (next phase)
scikit-learn>=1.1.0
tensorflow>=2.12.0  # or pytorch
xgboost>=1.6.0
prophet>=1.1.0
plotly>=5.0.0